@@ -17,55 +17,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.14",
-    "@angular/cdk": "^7.3.2",
-    "@angular/common": "7.2.14",
-    "@angular/compiler": "7.2.14",
-    "@angular/core": "7.2.14",
-    "@angular/fire": "^5.1.1",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "7.2.14",
-    "@angular/http": "7.2.14",
-    "@angular/material": "^7.3.2",
-    "@angular/platform-browser": "7.2.14",
-    "@angular/platform-browser-dynamic": "7.2.14",
-    "@angular/platform-server": "7.2.14",
-    "@angular/pwa": "^0.12.4",
-    "@angular/router": "7.2.14",
-    "@angular/service-worker": "7.2.14",
-    "core-js": "^2.5.3",
-    "firebase": "^5.8.2",
+    "@angular/animations": "8.2.4",
+    "@angular/cdk": "^8.1.4",
+    "@angular/common": "8.2.4",
+    "@angular/compiler": "8.2.4",
+    "@angular/core": "8.2.4",
+    "@angular/fire": "^5.2.1",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "8.2.4",
+    "@angular/http": "7.2.15",
+    "@angular/material": "^8.1.4",
+    "@angular/platform-browser": "8.2.4",
+    "@angular/platform-browser-dynamic": "8.2.4",
+    "@angular/platform-server": "8.2.4",
+    "@angular/pwa": "^0.803.2",
+    "@angular/router": "8.2.4",
+    "@angular/service-worker": "8.2.4",
+    "core-js": "^3.2.1",
+    "firebase": "^6.5.0",
     "hammerjs": "^2.0.8",
     "raf": "^3.4.0",
-    "rxjs": "^6.5.1",
-    "terser": "3.14.1",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "rxjs": "^6.5.2",
+    "terser": "4.2.1",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.6",
-    "@angular/cli": "7.3.8",
-    "@angular/compiler-cli": "7.2.14",
-    "@angular/language-service": "7.2.14",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "^0.803.2",
+    "@angular/cli": "8.3.2",
+    "@angular/compiler-cli": "8.2.4",
+    "@angular/language-service": "8.2.4",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~9.4.6",
-    "codelyzer": "~4.2.1",
-    "cypress": "^3.1.5",
-    "firebase-admin": "^5.9.0",
-    "firebase-functions": "^0.8.1",
-    "jasmine-core": "~3.1.0",
+    "@types/node": "~12.7.3",
+    "codelyzer": "~5.1.0",
+    "cypress": "^3.4.1",
+    "firebase-admin": "^8.4.0",
+    "firebase-functions": "^3.2.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.2",
-    "ts-node": "~5.0.0",
-    "tslint": "~5.9.1",
-    "typescript": "3.2.4",
-    "webpack-bundle-analyzer": "^2.10.0"
+    "ts-node": "~8.3.0",
+    "tslint": "~5.19.0",
+    "typescript": "3.5.3",
+    "webpack-bundle-analyzer": "^3.4.1"
   }
 }
